[URLReplacer] URL変換が有効になっていない場合でも元のメッセージが削除されてしまう問題を修正

+++ cogs/url_replacer.py
@@ -46,16 +46,17 @@ class CogURLReplacer(discord.Cog):
         if message.author.bot:
             return
         deleted = False
-        # vxtwitterとphixivのURLのみのとき
-        if re.fullmatch(rf'^(\s*({self.replacer_vxtwitter.url_pattern.pattern})|({self.replacer_phixiv.url_pattern.pattern})\s*)+$', message.content):
-            manage_messages = message.channel.permissions_for(message.guild.me)
-            # attachmentsがなく、マシロにメッセージ管理権限がある場合、元のメッセージを削除
-            if manage_messages and not message.attachments:
-                try:
-                    await message.delete()
-                    deleted = True
-                except discord.Forbidden:
-                    pass
+        if self.replacer_vxtwitter.is_enabled() or self.replacer_phixiv.is_enabled():
+            # vxtwitterとphixivのURLのみのとき
+            if re.fullmatch(rf'^(\s*({self.replacer_vxtwitter.url_pattern.pattern})|({self.replacer_phixiv.url_pattern.pattern})\s*)+$', message.content):
+                manage_messages = message.channel.permissions_for(message.guild.me)
+                # attachmentsがなく、マシロにメッセージ管理権限がある場合、元のメッセージを削除
+                if manage_messages and not message.attachments:
+                    try:
+                        await message.delete()
+                        deleted = True
+                    except discord.Forbidden:
+                        pass
 
         # 変換後のURLを送信
         await self.replacer_vxtwitter.send(message, deleted=deleted)
